+++ src/main/java/kitchenpos/product/repository/ProductRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.repository;
+package kitchenpos.product.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.domain.Product;
+import kitchenpos.product.model.Product;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
 }
