+++ src/test/java/kitchenpos/tablegroup/application/TableGroupServiceTest.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import kitchenpos.application.ServiceTest;
+import kitchenpos.ServiceTest;
 import kitchenpos.order.model.Order;
 import kitchenpos.order.model.OrderStatus;
 import kitchenpos.ordertable.model.OrderTable;
@@ -24,7 +24,6 @@ import kitchenpos.tablegroup.application.dto.TableGroupCreateRequestDto;
 import kitchenpos.tablegroup.application.dto.TableGroupResponseDto;
 import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.ordertable.repository.OrderTableRepository;
-import kitchenpos.tablegroup.application.TableGroupService;
 
 class TableGroupServiceTest extends ServiceTest {
     @Autowired
